Moved timezone-attribute from Event to Track
